+++ leap.yaml
@@ -4,14 +4,14 @@ secretId: ""
 entryFile: leap_binder.py
 include:
     - leap_binder.py
-    - coco_od/data/preprocessing.py
-    - coco_od/visualizers/visualizers.py
-    - coco_od/visualizers/visualizers_getters.py
-    - coco_od/utils/general_utils.py
-    - coco_od/utils/gcs_utils.py
-    - coco_od/utils/ioa_utils.py
-    - coco_od/yolo_helpers/yolo_utils.py
-    - coco_od/metrics.py
-    - coco_od/config.py
-    - coco_od/object_detection_config.yml
+    - yolo_od/data/preprocessing.py
+    - yolo_od/visualizers/visualizers.py
+    - yolo_od/visualizers/visualizers_getters.py
+    - yolo_od/utils/general_utils.py
+    - yolo_od/utils/gcs_utils.py
+    - yolo_od/utils/ioa_utils.py
+    - yolo_od/yolo_helpers/yolo_utils.py
+    - yolo_od/metrics.py
+    - yolo_od/config.py
+    - yolo_od/object_detection_config.yml
     - object_detection_config.yaml
