+++ yolo_od/utils/gcs_utils.py
@@ -8,7 +8,7 @@ from google.oauth2 import service_account
 
 from functools import lru_cache
 
-from coco_od.config import CONFIG
+from yolo_od.config import CONFIG
 
 
 @lru_cache()
